requirements.txt
# FastAPI y servidor ASGI
fastapi==0.116.1
uvicorn==0.35.0

# Manejo de retries para la API externa
tenacity==9.1.2

# Variables de entorno
python-dotenv==1.1.1

# Base de datos
SQLAlchemy==2.0.43

# Google GenAI
google-genai==1.32.0
genai==2.1.0
